requirements.txt
streamlit>=1.20
google-generativeai>=0.2.0
qdrant-client>=1.9.0
pdf2image>=1.16.0
pillow>=9.0.0
torch
numpy
colpali-engine
pytesseract
